User auth added

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../model/goalModel');
+const User = require('../model/userModel');
 
 //Desc Get all goals
 //Route GET /api/goals
 //Access Privet
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals);
 });
@@ -21,6 +22,7 @@ const createGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id
     });
 
     res.status(200).json(goal);
@@ -36,6 +38,21 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error('Goal not found');
     };
+
+    const user = await User.findById(req.user.id);
+
+    // Check user
+    if(!user) {
+        res.status(401);
+        throw new Error('User not fond');
+    };
+
+    // Chech if it's goal's user
+    if(goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
+    };
+
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -54,6 +71,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error(`Goal with id: ${req.params.id} Nof found`);
     };
+
+    const user = await User.findById(req.user.id);
+
+    // Check user
+    if (!user) {
+        res.status(401);
+        throw new Error('User not fond');
+    };
+
+    // Chech if it's goal's user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
+    };
+
 
     await goal.deleteOne();
     res.status(200).json({ message: `Goal with id: (${req.params.id}) has been deleted` });
